feat: game_room과 board_piece 저장을 위한 JDBC DAO를 구현
GameRoom, BoardPiece DTO를 DB 저장 구조에 맞게 정리
ConnectionManager를 추가해 JDBC 연결 책임 분리
GameRoomDao, BoardPieceDao 인터페이스 추가
JdbcGameRoomDao, JdbcBoardPieceDao를 구현해 game_room / board_piece CRUD 기반 마련

+++ src/main/java/janggi/dto/BoardPiece.java
@@ -1,6 +1,6 @@
 package janggi.dto;
 
-public record BoardPiece(int gameRoomId,
+public record BoardPiece(long gameRoomId,
                          int rowPos,
                          int colPos,
                          String pieceType,
